packages/python-sdk/pyproject.toml
[project]
name = "relayauth"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["httpx>=0.25", "PyJWT>=2.8", "cryptography>=41"]

[project.optional-dependencies]
dev = ["pytest>=7", "pytest-asyncio>=0.21", "respx>=0.20"]
